refactor: took out checkboxButton reducer to redux store
