+++ client/src/components/health-log-tab/DiagnosticContainer.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import { loadObservations } from '../redux/actions/observations';
+import { loadObservations } from '../../redux/actions/observations';
 import { useDispatch } from 'react-redux';
 import HealthLogToggle from './HealthLogToggle';
-import Scroll from './Scroll'
+import Scroll from '../Scroll'
 
 const dateOptions = {
   weekday: 'long',
@@ -12,7 +12,7 @@ const dateOptions = {
 };
 
 function DiagnosticContainer(props) {
-  const { userSession} = props;
+  const { userSession } = props;
   const dispatch = useDispatch();
   const today = new Date();
   dispatch(loadObservations());
@@ -20,12 +20,12 @@ function DiagnosticContainer(props) {
     <div className="DiagnosticContainer">
 
       <h4>
-              Hello, <b>{userSession.loadUserData().profile.name} </b>
-            </h4>
-            <h5>
-              Today is <b>{today.toLocaleDateString(undefined, dateOptions)}</b>{' '}
-            </h5>
-            <hr className="hr" />
+        Hello, <b>{userSession.loadUserData().profile.name} </b>
+      </h4>
+      <h5>
+        Today is <b>{today.toLocaleDateString(undefined, dateOptions)}</b>{' '}
+      </h5>
+      <hr className="hr" />
 
       {/*uncomment out below to show popup*/}
       {/* <Container>
