+++ client/src/components/health-log-tab/HealthLogToggle.js
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import ButtonGroup from '@material-ui/core/ButtonGroup';
 import Button from '@material-ui/core/Button';
 import Container from '@material-ui/core/Container';
-import MyHealthLog from './MyHealthLog';
+import MyHealthLog from './my-health-log/MyHealthLog';
 
 export default function HealthLogToggle() {
   const [toggleValue, setToggleValue] = useState('myHealthLog');
