+++ client/src/redux/constants.js
@@ -1,0 +1,6 @@
+// ACTION TYPES
+export const HANDLE_CHECKBOX_CHANGE = 'HANDLE_CHECKBOX_CHANGE';
+export const HANDLE_BUTTON_GROUP_CHANGE = 'HANDLE_BUTTON_GROUP_CHANGE';
+export const HANDLE_FEVER = 'HANDLE_FEVER';
+export const HANDLE_HOURS = 'HANDLE_HOURS';
+export const HANDLE_MINUTES = 'HANDLE_MINUTES';
